feat: code-splitting new search for reduced initial js load
