+++ src/components/MainLayout/Header/SearchDropdown/index.ts
@@ -1,0 +1,3 @@
+import SearchDropdown from "./SearchDropdown";
+
+export default SearchDropdown;
