+++ src/utils/constants.ts
@@ -1,3 +1,5 @@
+import { SearchFilterTypes } from "@schema/search.schema";
+
 const environmentUrl =
   process.env.NEXT_PUBLIC_BYPASS_URL || process.env.NEXT_PUBLIC_VERCEL_URL;
 
@@ -6,3 +8,10 @@ export const baseUrl = environmentUrl
   : `http://localhost:3000`;
 
 export const url = `${baseUrl}/api/trpc`;
+
+export const SEARCH_FILTERS: SearchFilterTypes[] = [
+  "posts",
+  "comments",
+  "tags",
+  "users",
+];
